AllShapes component and NOT WORKING GAME
how did this turn out so bad. This was so so bad the app just does NOT work at all. This game is failing so bad and none of it is functioning properly. The API key needs to be refreshed because the first one was pushed. Game is game tho