Coding finished
although there is still so much to improve on. I am overall happy with this product and feel confident delivering it. Maybe if I go back to it I might add a way for the user to access activity from the mobile section but other than that everything is looking great :D